requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
discord.py==1.1.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauthlib==3.0.1
PySocks==1.6.8
python-dotenv==0.10.1
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
toml==0.10.0
appdirs==1.4.3
black==18.9b0
Click==7.0
filelock==3.0.10
jsonschema==3.0.1
pluggy==0.9.0
py==1.8.0
tox==3.9.0
virtualenv==16.4.3